util.scm: Initialize
The goal is to reduce code duplication between readit.scm and
readit-graph.scm. For now, the file only defines utility procedures for
parsing input files.

From my current understanding, it seems that utility code is also
compiled twice. As the file is very small at the moment, this shouldn't
be a problem but long term it might be desirable to convert it to a
module.